pyproject.toml
[project]
authors = [{email = "seria.ati@gmail.com", name = "seriaati"}]
dependencies = [
  "aiohttp-client-cache[sqlite]>=0.12.4",
  "fastapi>=0.114.0",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.30.6",
]
description = "Fix twitch clip Discord embeds."
license = {file = "LICENSE"}
name = "fxtwitch"
readme = "README.md"
requires-python = ">=3.11"
version = "0.2.0"

[tool.pyright]
typeCheckingMode = "standard"
